docs: remove fleet overview docs and fix test suite for instance-admin access model
